[osx][windowing] relocate SDL specific code to an SDL folder
relocates SDL implementations to a specific SDL folder
