+++ xbmc/windowing/osx/CMakeLists.txt
@@ -1,18 +1,9 @@
 set(SOURCES CocoaDPMSSupport.cpp
             OSScreenSaverOSX.cpp
-            WinEventsOSX.mm
-            WinSystemOSX.mm
             VideoSyncOsx.mm)
 set(HEADERS CocoaDPMSSupport.h
             OSScreenSaverOSX.h
-            WinEventsOSX.h
-            WinSystemOSX.h
             VideoSyncOsx.h)
-
-if(SDL_FOUND)
-  list(APPEND SOURCES WinEventsSDL.cpp)
-  list(APPEND HEADERS WinEventsSDL.h)
-endif()
 
 if(OPENGL_FOUND)
   list(APPEND SOURCES WinSystemOSXGL.mm)
