+++ xbmc/windowing/osx/SDL/WinSystemOSXSDL.mm
@@ -6,14 +6,11 @@
  *  See LICENSES/README.md for more information.
  */
 
-#include "WinSystemOSX.h"
+#include "WinSystemOSXSDL.h"
 
 #include "AppInboundProtocol.h"
 #include "CompileInfo.h"
-#include "OSScreenSaverOSX.h"
 #include "ServiceBroker.h"
-#include "VideoSyncOsx.h"
-#include "WinEventsOSX.h"
 #include "cores/AudioEngine/AESinkFactory.h"
 #include "cores/AudioEngine/Sinks/AESinkDARWINOSX.h"
 #include "cores/RetroPlayer/process/osx/RPProcessInfoOSX.h"
@@ -37,6 +34,9 @@
 #include "utils/SystemInfo.h"
 #include "utils/log.h"
 #include "windowing/osx/CocoaDPMSSupport.h"
+#include "windowing/osx/OSScreenSaverOSX.h"
+#include "windowing/osx/SDL/WinEventsSDL.h"
+#include "windowing/osx/VideoSyncOsx.h"
 
 #include "platform/darwin/DarwinUtils.h"
 #include "platform/darwin/DictionaryUtils.h"
